requirements.txt
python-dotenv
fastapi
google-ai-generativelanguage
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-vision
google-genai
google-generativeai
googleapis-common-protos
groq
langchain-core
numpy
pandas
pillow
psycopg2-binary
pydantic
requests
streamlit
uvicorn
